More descriptive output & Bugfixes
The output of the program is now more descriptive, and provides more debugging information...mostly for me...
Fixed bugs with default values being set to zero, and added a case for the timing vector where the user inputs nothing, and it uses the default values.
There are more comments in the code
There are two png files that document the shifting of my function on simulated data.  Also mostly for me.